rename public gem to mx-atrium
