--- lib/atrium/credential.rb
@@ -1,11 +1,0 @@
-module Atrium
-  class Credential
-    include ::ActiveAttr::Model
-
-    attribute :field_name
-    attribute :guid
-    attribute :label
-    attribute :type
-    attribute :value
-  end
-end
