--- lib/atrium/error.rb
@@ -1,4 +1,0 @@
-module Atrium
-  class Error < StandardError
-  end
-end
