--- lib/atrium/version.rb
@@ -1,3 +1,0 @@
-module Atrium
-  VERSION = "0.2.3"
-end
