+++ lib/mx/atrium/error.rb
@@ -1,0 +1,6 @@
+module MX
+  module Atrium
+    class Error < StandardError
+    end
+  end
+end
