+++ lib/mx/atrium/version.rb
@@ -1,0 +1,5 @@
+module MX
+  module Atrium
+    VERSION = "0.2.3"
+  end
+end
